requirements.txt
pillow
matplotlib
jupyter
ipykernel
numpy
pandas
scipy
tqdm
flake8
autopep8
opencv-python